Improve handling and report when cleaning repos

+++ diamond.releng.jenkins/job.scripts/update_single_git_repo_function.sh
@@ -100,9 +100,11 @@ update_single_git_repo_function () {
     # We cannot assume that the local branch is tracking the standard branch, since this might be a Gerrit repository previously used to test a change.
     set -e  # Turn errexit back on
     git -C ${repo_path} checkout --detach --quiet
-    git -C ${repo_path} branch -D --quiet ${repo_branch} |& sed "s/^/[${repo_name}] /" || true
+    if git show-ref --verify --quiet "refs/heads/${repo_branch}"; then
+        git -C ${repo_path} branch -D --quiet ${repo_branch} |& sed "s/^/[${repo_name}] /" || true
+    fi
     git -C ${repo_path} checkout -b ${repo_branch} remotes/origin/${repo_branch} --no-track --quiet |& sed "s/^/[${repo_name}] /"
-    echo "Clean and reset of ${repo_name} completed"
+    echo "Clean and reset of ${repo_name} to branch ${repo_branch} completed"
 
     $([ "$oldpipefail" == "0" ]) && set -o pipefail || true  # Turn pipefail on if it was on at the top of this script
     $([ "$oldpipefail" == "1" ]) && set +o pipefail || true  # Turn pipefail off if it was off at the top of this script
